+++ src/imputation/imputation_main.py
@@ -10,82 +10,9 @@ from src.imputation.short_to_long import run_short_to_long
 from src.imputation.MoR import run_mor
 from src.imputation.sf_expansion import run_sf_expansion
 from src.imputation import tmi_imputation as tmi
+from src.imputation import manual_imputation as mimp
 
 ImputationMainLogger = logging.getLogger(__name__)
-
-
-def add_trim_column(df: pd.DataFrame, column_name: str = 'manual_trim', value: bool = False) -> pd.DataFrame:
-    """
-    Adds a new column to a DataFrame with a default value.
-
-    Args:
-        df (pd.DataFrame): The DataFrame to add the new column to.
-        column_name (str, optional): The name of the new column. Defaults to 'manual_trim'.
-        value (bool, optional): The default value for the new column. Defaults to False.
-
-    Returns:
-        pd.DataFrame: The DataFrame with the new column added.
-
-    Raises:
-        ValueError: If the DataFrame is empty or the column already exists in the DataFrame.
-    """
-    if df.empty:
-        raise ValueError("The DataFrame is empty.")
-    
-    if column_name in df.columns:
-        raise ValueError(f"A column with name {column_name} already exists in the DataFrame.")
-
-    df[column_name] = value
-    return df
-
-
-def get_latest_csv(directory: str) -> str:
-    """
-    Gets the latest CSV file in a directory.
-
-    Args:
-        directory (str): The directory to search for CSV files.
-
-    Returns:
-        str: The path of the latest CSV file, or an empty string if no CSV files are found.
-    """
-    list_of_files = glob.glob(os.path.join(directory, "*.csv"))  # get list of all csv files
-    if not list_of_files:  # if list is empty, return empty string
-        return None
-    latest_file = max(list_of_files, key=os.path.getctime) # get the latest file
-    
-    return latest_file
-
-# check if any files are in imputation/manual_trimming and check if load_manual_imputation is True
-# if so load the file and any records which are marked True in the manual_trim column will be 
-# excluded from the imputation process and will be output as is. They will be marked as 'manual_trim' in the imp_marker column
-def load_manual_imputation(config: Dict[str, Any], 
-                           df: pd.DataFrame,
-                           isfile_func: callable) -> pd.DataFrame:
-    """
-    Loads a manual trimming file if it exists and adds a manual_trim column to the DataFrame.
-
-    Args:
-        config (Dict[str, Any]): The configuration dictionary.
-        df (pd.DataFrame): The dataframe to be imputed.
-        isfile_func (callable): The function to use to check if the file exists.
-    Returns:
-        pd.DataFrame: The DataFrame with the manual_trim column added.
-    """
-    NETWORK_OR_HDFS = config["global"]["network_or_hdfs"]
-    imp_path = config[f"{NETWORK_OR_HDFS}_paths"]["imputation_path"]
-
-    new_man_trim_file = get_latest_csv(f"{imp_path}/manual_trimming/")
-    
-    if config["global"]["load_manual_imputation"] and isfile_func(new_man_trim_file):
-        manual_trim_df = pd.read_csv(new_man_trim_file)
-        df = df.merge(manual_trim_df, on=["reference", "instance"], how="left")
-        df = df.drop(columns=["manual_trim_x"])
-        df = df.rename(columns={"manual_trim_y": "manual_trim"})
-        df["manual_trim"] = df["manual_trim"].fillna(False)
-    else:
-        df = add_trim_column(df)
-    return df
 
 
 def run_imputation(
@@ -105,7 +32,7 @@ def run_imputation(
 
     # Apportion cols 4xx and 5xx to create FTE and headcount values
     df = run_apportionment(df)
-        
+
     # Convert shortform responses to longform format
     df = run_short_to_long(df)
 
@@ -149,7 +76,7 @@ def run_imputation(
     imp_path = config[f"{NETWORK_OR_HDFS}_paths"]["imputation_path"]
 
     # Add a manual_trim column to the QA df
-    qa_df = add_column(qa_df, column_name="manual_trim", value=False)
+    qa_df = mimp.add_trim_column(qa_df, column_name="manual_trim", value=False)
 
     if config["global"]["output_imputation_qa"]:
         ImputationMainLogger.info("Outputting Imputation files.")
